30/10/2023
Ended the part 1 definition & implementation of the class but we do have to check the Getstrike methods in order to do it i retrieve it from a public methods that i wanted first as protected in my class VanillaOption. In fact if i want to do it i have in the .H file of VanillaOption to assign the BlackScholesPricer class as a friend.